post controller

+++ server/controllers/postController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require('express-async-handler');
 
 const Post = require('../models/postModel');
+const User = require('../models/userModel');
 
 // @desc    Get posts
 // @route   GET /api/posts
-// @access  Private
+// @access  Public
 const getPosts = asyncHandler(async (req, res) => {
     const posts = await Post.find();
 
@@ -24,7 +25,8 @@ const setPost = asyncHandler(async (req, res) => {
     const post = await Post.create({
         title: req.body.title,
         snippet: req.body.snippet,
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(post);
@@ -41,6 +43,18 @@ const updatePost = asyncHandler(async (req, res) => {
         throw new Error('Post not found');
     }
 
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    if (post.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
+    }
+
     const updatedPost = await Post.findByIdAndUpdate(req.params.id, req.body, { new: true });
 
     res.status(200).json(updatedPost);
@@ -55,6 +69,18 @@ const deletePost = asyncHandler(async (req, res) => {
     if (!post) {
         res.status(400);
         throw new Error('Post not found');
+    }
+
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    if (post.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     const deletedPost = await Post.findByIdAndDelete(req.params.id);
